Ceci est une démonstration 🪽
French Translation: This is a demonstration.

I will be using this design to test the routes when interacting with the Self-Custody Wallets and Multi-Signature Wallets. It will be minted on Songbird, a Canary Network, where smart contracts can be developed for deployment while being interconnected with other blockchains that have real-world value.

Why am I doing this?
To ensure the platform on NODEWEAR works as designed, and to show transparency of the infrastructure.

Who is this for?
Artists who create digital art who wants to market their work utilizing blockchain.

This design will be used during the new upcoming cohort by @atlantachain when fine tuning the platform with feedback from peers and other developers.